message item and message empty component
