+++ App.js
@@ -109,42 +109,8 @@ const App = () => {
 
   useEffect(() => {
     (async () => {
-      // io = await socketClient.init();
-
-      // io.on("connect", () => {
-      //   console.log('socket connected');
-      // });
-
-      // io.on('connect_error', socket => {
-      //   console.log(`socket connect error ===> ${socket}`);
-      // });
-
       // ask for notification persmission
       await askForPermission()
-      // const channelId = await notifee.createChannel({
-      //   id: 'default',
-      //   name: 'Default Channel',
-      // });
-
-      // // Heartbeat.updateNotification({
-      // //   channelId,
-      // //   notificationId: 'widget-notification',
-      // // })
-
-      // setTimeout(async () => {
-      //   await notifee.displayNotification({
-      //     title: 'Welcome to VoIP Suite',
-      //     body: 'Main body content of the notification',
-      //     android: {
-      //       channelId,
-      //       // smallIcon: 'name-of-a-small-icon', // optional, defaults to 'ic_launcher'.
-      //       // pressAction is needed if you want the notification to open the app when pressed
-      //       pressAction: {
-      //         id: 'default',
-      //       },
-      //     },
-      //   });
-      // }, 5000);
     })();
 
     return () => {
