+++ index.js
@@ -7,9 +7,10 @@ import 'react-native-gesture-handler';
 import { AppRegistry } from 'react-native';
 import App from './App';
 import { name as appName } from './app.json';
-import { setupNotifeeHandlers } from './helpers/notifee';
+import { createChannel, displayNotification, setupNotifeeHandlers } from './helpers/notifee';
 import notifee from '@notifee/react-native';
 import socketClient from './helpers/socket';
+import { MESSAGE_CHANNEL_ID, MESSAGE_CHANNEL_NAME } from './helpers/config';
 
 const MyHeadlessTask = async () => {
   console.log('Receiving HeartBeat!');
@@ -35,28 +36,12 @@ const MyHeadlessTask = async () => {
 
     io.emit("join_profile_channel", '621e9f2685a90200160c3160');
     io.on("user_message", async function (data) {
-      console.log('data ===>', data)
+      console.log('data notif ===>', data)
       const { number, message } = data;
-      const channelId = await notifee.createChannel({
-        id: 'w-channel',
-        name: 'Test w-channel',
-      });
 
-      await notifee.displayNotification({
-        title: message,
-        body: 'Main body content of the notification',
-        data: { 'event_type': 'message', message: data },
-        android: {
-          channelId,
-          pressAction: {
-            id: 'default',
-          },
-        },
-      });
-
-      // init()
+      const channelId = await createChannel(MESSAGE_CHANNEL_ID, MESSAGE_CHANNEL_NAME)
+      await displayNotification(message, channelId, 'Main body content of the notification', { 'event_type': 'message', message: data })
     });
-
   }
 };
